Challenge Functional programming
Solutions of challenges for this secction where in the calculator challenge print the expressions using random values for each operation.

There is a class called Streams.jav where is resolve the challenge for streams using reduce and duplicate.

Finally the solution for the problem of office->table->element is resolve in the project